requirements.txt
torch>=1.13.0     
opencv-python>=4.6.0.66
Pillow>=9.3.0
numpy>=1.21.6
timm==0.6.12
ultralytics>=8.0.0
openvino>=2023.0.0
tqdm>=4.64.1
matplotlib>=3.7.0
scipy>=1.10.0
